add speedrun api client package
